Psr0Fixer - moved to contrib level
