+++ Symfony/CS/Console/Command/FixCommand.php
@@ -107,7 +107,7 @@ final class FixCommand extends Command
                     new InputOption('config', '', InputOption::VALUE_REQUIRED, 'The configuration name', null),
                     new InputOption('config-file', '', InputOption::VALUE_OPTIONAL, 'The path to a .php_cs file ', null),
                     new InputOption('dry-run', '', InputOption::VALUE_NONE, 'Only shows which files would have been modified'),
-                    new InputOption('level', '', InputOption::VALUE_REQUIRED, 'The level of fixes (can be psr0, psr1, psr2, or symfony (formerly all))', null),
+                    new InputOption('level', '', InputOption::VALUE_REQUIRED, 'The level of fixes (can be psr1, psr2, or symfony (formerly all))', null),
                     new InputOption('using-cache', '', InputOption::VALUE_REQUIRED, 'Does cache should be used (can be yes or no)', null),
                     new InputOption('cache-file', '', InputOption::VALUE_REQUIRED, 'The path to the cache file'),
                     new InputOption('fixers', '', InputOption::VALUE_REQUIRED, 'A list of fixers to run'),
@@ -130,7 +130,6 @@ The <comment>--verbose</comment> option will show the applied fixers. When using
 The <comment>--level</comment> option limits the fixers to apply on the
 project:
 
-    <info>php %command.full_name% /path/to/project --level=psr0</info>
     <info>php %command.full_name% /path/to/project --level=psr1</info>
     <info>php %command.full_name% /path/to/project --level=psr2</info>
     <info>php %command.full_name% /path/to/project --level=symfony</info>
@@ -223,7 +222,7 @@ then specify all fixers to be used:
     ?>
 
 You may also use a blacklist for the Fixers instead of the above shown whitelist approach.
-The following example shows how to use all ``symfony`` Fixers but the ``psr0`` fixer.
+The following example shows how to use all ``symfony`` Fixers but the ``short_tag`` fixer.
 Note the additional <comment>-</comment> in front of the Fixer name.
 
     <?php
@@ -234,7 +233,7 @@ Note the additional <comment>-</comment> in front of the Fixer name.
     ;
 
     return Symfony\CS\Config\Config::create()
-        ->fixers(array('-psr0'))
+        ->fixers(array('-short_tag'))
         ->finder(\$finder)
     ;
 
@@ -253,7 +252,7 @@ The ``psr2`` level is set by default, you can also change the default level:
 In combination with these config and command line options, you can choose various usage.
 
 For example, the default level is ``psr2``, but if you don't want to use
-the ``psr0`` fixer, you can specify the ``--fixers="-psr0"`` option.
+the ``short_tag`` fixer, you can specify the ``--fixers="-short_tag"`` option.
 
 But if you use the ``--fixers`` option with only exact fixers,
 only those exact fixers are enabled whether or not level is set.
