+++ Symfony/CS/Console/ConfigurationResolver.php
@@ -280,7 +280,6 @@ final class ConfigurationResolver
     {
         static $levelMap = array(
             'none' => FixerInterface::NONE_LEVEL,
-            'psr0' => FixerInterface::PSR0_LEVEL,
             'psr1' => FixerInterface::PSR1_LEVEL,
             'psr2' => FixerInterface::PSR2_LEVEL,
             'symfony' => FixerInterface::SYMFONY_LEVEL,
